@@ -1,7 +1,7 @@
 /* #undef HAVE_MAKE_UNIQUE */
 /* #undef HAVE_CLAMP */
 /* #undef USE_ISPC */
-/* #undef OS_WINDOWS */
+#define OS_WINDOWS
 /* #undef OS_LINUX */
 /* #undef OS_BSD */
-#define OS_MACOS
+/* #undef OS_MACOS */
